+++ modules/core/src/main/java/org/apache/ignite/internal/util/IgniteUtils.java
@@ -12536,4 +12536,9 @@ public abstract class IgniteUtils {
 
         return sb.toString();
     }
+
+    /** */
+    public static boolean isTxAwareQueriesEnabled(GridKernalContext kctx) {
+        return kctx.config().getTransactionConfiguration().isTxAwareQueriesEnabled();
+    }
 }
